movies_list/requirements.txt
tmdbsimple
django
djangorestframework